Move Indent and Shape to shape.rs from lib.rs
